requirements.txt
fastapi
uvicorn[standard]
pydantic
openvino
optimum
optimum-intel
huggingface_hub
transformers
torch
numpy
soundfile
librosa
pretty_midi
mido
pyfluidsynth
python-dotenv
requests
orjson
pyyaml
onnx
ruff
black
mypy
pytest
httpx
types-requests
types-PyYAML
